moved psync_delete_cached_crypto_keys to psynclib
